+++ tests/testthat/test-getdaylighthours.R
@@ -1,4 +1,4 @@
-context("test-goesaodc_getdaylighthours")
+context("test-getdaylighthours")
 
 test_that("Correct daylight sunrise and sunset hours are calculated", {
   
@@ -15,28 +15,28 @@ test_that("Correct daylight sunrise and sunset hours are calculated", {
                                             "America/Los_Angeles")
   
   # ---- GET DAYLIGHT HOURS WITH LAT/LON ---------------------------------------
-  ll_test_data <- goesaodc_getDaylightHours(datetime = "2019-09-06", 
-                                            longitude = -123.245, 
-                                            latitude = 42.861)
+  ll_test_data <- getDaylightHours(datetime = "2019-09-06", 
+                                   longitude = -123.245, 
+                                   latitude = 42.861)
   
   # ---- GET DAYLIGHT HOURS WITH BBOX ------------------------------------------
-  bb_test_data <- goesaodc_getDaylightHours(datetime = "2019-09-06", 
-                                            bbox = c(-124.566, -116.463, 
-                                                     41.991, 46.292))
+  bb_test_data <- getDaylightHours(datetime = "2019-09-06", 
+                                   bbox = c(-124.566, -116.463, 
+                                            41.991, 46.292))
   
   # ---- GET DAYLIGHT HOURS WITH DATETIME AND TIMEZONE -------------------------
-  tz_test_data <- goesaodc_getDaylightHours(datetime = "2019-09-06", 
-                                            timezone = "America/Los_Angeles")
+  tz_test_data <- getDaylightHours(datetime = "2019-09-06", 
+                                   timezone = "America/Los_Angeles")
   
   # --- GET DAYLIGHT HOURS WITH POSIXt ALONE -----------------------------------
-  posix_test_data <- goesaodc_getDaylightHours(posix_t)
+  posix_test_data <- getDaylightHours(posix_t)
   
   # ---- LAT/LON TIME DIFFS ----------------------------------------------------
   ll_sunrise_diff <- as.numeric(difftime(ll_dayInfo$sunrise,
                                          ll_test_data$sunrise,
                                          units = "mins"))
-                                         
-                                
+  
+  
   ll_sunset_diff <- as.numeric(difftime(ll_dayInfo$sunset, 
                                         ll_test_data$sunset, 
                                         units = "mins"))
@@ -47,8 +47,8 @@ test_that("Correct daylight sunrise and sunset hours are calculated", {
                                          units = "mins"))
   
   bb_sunset_diff <- as.numeric(difftime(ll_dayInfo$sunset,
-                                         ll_test_data$sunset,
-                                         units = "mins"))
+                                        ll_test_data$sunset,
+                                        units = "mins"))
   
   # ---- TZ TIME DIFFS ---------------------------------------------------------
   tz_sunrise_diff <- as.numeric(difftime(ll_dayInfo$sunrise,
@@ -56,13 +56,13 @@ test_that("Correct daylight sunrise and sunset hours are calculated", {
                                          units = "mins"))
   
   tz_sunset_diff <- as.numeric(difftime(ll_dayInfo$sunset,
-                                         tz_test_data$sunset,
-                                         units = "mins"))
-
+                                        tz_test_data$sunset,
+                                        units = "mins"))
+  
   # ---- POSIX TIME DIFF (SHOULD BE EQUAL TO TZ TIME ---------------------------                                         
   posix_sunrise_diff <- as.numeric(difftime(tz_test_data$sunrise,
-                                           posix_test_data$sunrise,
-                                           units = "secs"))
+                                            posix_test_data$sunrise,
+                                            units = "secs"))
   
   posix_sunset_diff <- as.numeric(difftime(tz_test_data$sunset,
                                            posix_test_data$sunset,
@@ -88,21 +88,27 @@ test_that("function fails when passed incorrect parameters", {
   
   # No datetime given
   expect_error(
-    goesaodc_getDaylightHours(),
+    getDaylightHours(),
     regexp = "argument 'datetime' must not be NULL."
-  )
-  
-  # Too much geographic data given
-  expect_error(
-    goesaodc_getDaylightHours(datetime = "2019-09-06", 
-                              bbox = c(0,0,0,0), 
-                              latitude = 0, 
-                              longitude = 0),
-    regexp = "Cannot process both BBOX and Lat/Lon at the same time. Pick one."
   )
   
   # No data available to derive timezone
   expect_error(
-    goesaodc_getDaylightHours(datetime = "2019-09-06"),
-    regexp = "Can't calculate timezone. Need timezone, lat/lon, or  bbox.")
+    getDaylightHours(datetime = "2019-09-06")
+  )
+  
+})
+
+test_that("function is vectorized", {
+  
+  daylight <- getDaylightHours(
+    datetime = c("2019-06-21", "2019-09-23", "2019-12-22"),
+    longitude = -123.245, 
+    latitude = 42.861
+  )
+  
+  expect_true(lubridate::is.POSIXt(daylight$sunrise))
+  expect_equal(length(daylight$sunrise), 3)
+  expect_equal(strftime(daylight$sunrise, "%Z"), c("PDT", "PDT", "PST"))
+  
 })